requirements.txt
pyspark~=3.5.0
python-dotenv~=1.0.0
